requirements.txt
Django==6.0.3
djangorestframework==3.17.1
psycopg2-binary==2.9.11
celery==5.6.3
redis==7.4.0
django-ratelimit==4.1.0
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-debug-toolbar==6.2.0
channels==4.3.2
channels-redis==4.3.0
daphne==4.1.2
django-fsm==3.0.1
pytest-django==4.12.0
pytest-cov==4.1.0
django-extensions==4.1
locust==2.43.3
prometheus-client==0.24.1
python-decouple==3.8
django-environ==0.10.0
black==23.12.1
isort==5.13.2
flake8==6.1.0
safety==2.3.5
bandit==1.7.6
